refactor(cli): derive some defaults and extract retry
