+++ cli/src/cmd/forge/debug.rs
@@ -1,8 +1,5 @@
-use super::{build::BuildArgs, script::ScriptArgs, watch::WatchArgs};
-use crate::{
-    cmd::forge::{build::CoreBuildArgs, create::RETRY_VERIFY_ON_CREATE},
-    opts::MultiWallet,
-};
+use super::{build::BuildArgs, script::ScriptArgs};
+use crate::cmd::{forge::build::CoreBuildArgs, retry::RETRY_VERIFY_ON_CREATE};
 use clap::{Parser, ValueHint};
 use foundry_common::evm::EvmArgs;
 use std::path::PathBuf;
@@ -49,26 +46,12 @@ impl DebugArgs {
             args: self.args,
             target_contract: self.target_contract,
             sig: self.sig,
-            legacy: false,
-            broadcast: false,
-            skip_simulation: false,
             gas_estimate_multiplier: 130,
-            opts: BuildArgs {
-                args: self.opts,
-                names: false,
-                sizes: false,
-                watch: WatchArgs::default(),
-            },
-            wallets: MultiWallet::default(),
+            opts: BuildArgs { args: self.opts, ..Default::default() },
             evm_opts: self.evm_opts,
-            resume: false,
             debug: true,
-            slow: false,
-            etherscan_api_key: None,
-            verify: false,
-            json: false,
-            with_gas_price: None,
             retry: RETRY_VERIFY_ON_CREATE,
+            ..Default::default()
         };
         script.run_script().await
     }
