+++ cli/src/cmd/mod.rs
@@ -9,5 +9,7 @@ pub mod cast;
 pub mod forge;
 
 // Re-export our shared utilities
+mod retry;
 mod utils;
+
 pub use utils::*;
